% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_suite.R
\name{bayesian_meta_analysis}
\alias{bayesian_meta_analysis}
\title{Bayesian Meta-Analysis with Prior Elicitation}
\usage{
bayesian_meta_analysis(
  data,
  priors = NULL,
  method = c("mcmc", "empirical"),
  chains = 4,
  iter = 4000,
  warmup = 1000
)
}
\arguments{
\item{data}{Meta-analysis data}

\item{priors}{List of prior specifications}

\item{method}{Bayesian method: "mcmc", "empirical"}

\item{chains}{Number of MCMC chains}

\item{iter}{Number of iterations}

\item{warmup}{Warmup iterations}
}
\description{
Bayesian Meta-Analysis with Prior Elicitation
}
