% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_fix.R
\name{bayesian_meta_analysis_fixed}
\alias{bayesian_meta_analysis_fixed}
\title{Bayesian Meta-Analysis Wrapper}
\usage{
bayesian_meta_analysis_fixed(data, ...)
}
\description{
Bayesian Meta-Analysis Wrapper
}
