% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_meta_analysis.R
\name{calculate_heterogeneity_statistics}
\alias{calculate_heterogeneity_statistics}
\title{Calculate Heterogeneity Statistics}
\usage{
calculate_heterogeneity_statistics(data, meta_result)
}
\arguments{
\item{data}{Current data subset}

\item{meta_result}{Meta-analysis results}
}
\value{
Heterogeneity statistics
}
\description{
Calculates I-squared, H-squared, and Q statistics
}
