% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_meta_analysis.R
\name{calculate_pairwise_comparisons}
\alias{calculate_pairwise_comparisons}
\title{Calculate Pairwise Treatment Comparisons}
\usage{
calculate_pairwise_comparisons(treatment_effects)
}
\arguments{
\item{treatment_effects}{Treatment effect estimates}
}
\value{
Data frame of pairwise comparisons
}
\description{
Calculates all possible pairwise treatment comparisons
}
