% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_interval_fix.R
\name{calculate_prediction_interval_fixed}
\alias{calculate_prediction_interval_fixed}
\title{Calculate Prediction Interval Wrapper}
\usage{
calculate_prediction_interval_fixed(result, level = 0.95)
}
\description{
Calculate Prediction Interval Wrapper
}
