% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_meta_analysis.R
\name{calculate_treatment_rankings}
\alias{calculate_treatment_rankings}
\title{Calculate Treatment Rankings}
\usage{
calculate_treatment_rankings(nma_results)
}
\arguments{
\item{nma_results}{Network meta-analysis results}
}
\value{
Treatment ranking results
}
\description{
Calculates treatment ranking probabilities
}
