% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm-package.R
\docType{package}
\name{cbamm-package}
\alias{cbamm-package}
\title{cbamm: Comprehensive Bayesian and Advanced Meta-Analysis Methods}
\description{
The cbamm package provides tools for conducting comprehensive meta-analyses
using both frequentist and Bayesian approaches.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cbamm-dev/cbamm}
  \item Report bugs at \url{https://github.com/cbamm-dev/cbamm/issues}
}

}
\author{
\strong{Maintainer}: Mahmood Ahmad \email{mahmood.ahmad@ahmadiyya.org} (Ahmadiyya Muslim Community)

}
\keyword{internal}
