% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{compute_analysis_weights}
\alias{compute_analysis_weights}
\title{Compute Analysis Weights from Transport Weights}
\usage{
compute_analysis_weights(data, transport_weights = NULL)
}
\arguments{
\item{data}{Study data}

\item{transport_weights}{Transport weights (optional)}
}
\value{
Vector of analysis weights
}
\description{
Compute Analysis Weights from Transport Weights
}
