% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_meta_analysis.R
\name{create_contrast_data}
\alias{create_contrast_data}
\title{Create Contrast Data for Network Meta-Analysis}
\usage{
create_contrast_data(network_data, reference)
}
\arguments{
\item{network_data}{Prepared network data}

\item{reference}{Reference treatment}
}
\value{
Contrast data frame
}
\description{
Converts network data to treatment contrast format
}
