% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_meta_analysis.R
\name{create_vcov_matrix}
\alias{create_vcov_matrix}
\title{Create Variance-Covariance Matrix}
\usage{
create_vcov_matrix(contrast_data)
}
\arguments{
\item{contrast_data}{Contrast data frame}
}
\value{
Variance-covariance matrix
}
\description{
Creates variance-covariance matrix accounting for within-study correlations
}
