% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_design_wrapper.R
\name{cross_design_synthesis_unified}
\alias{cross_design_synthesis_unified}
\title{Cross-Design Synthesis Wrapper}
\usage{
cross_design_synthesis_unified(data, design_var = "design", ...)
}
\description{
Cross-Design Synthesis Wrapper
}
