% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{.compute_entropy_weights}
\alias{.compute_entropy_weights}
\title{Compute Entropy-based Transport Weights}
\usage{
.compute_entropy_weights(X, target_moments, truncation, data)
}
\description{
Compute Entropy-based Transport Weights
}
\keyword{internal}
