% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_meta_analysis.R
\name{extract_treatment_pairs}
\alias{extract_treatment_pairs}
\title{Extract Treatment Pairs from Studies}
\usage{
extract_treatment_pairs(data, studies, treatments)
}
\arguments{
\item{data}{Input data frame}

\item{studies}{Study identifier column name}

\item{treatments}{Treatment identifier column name}
}
\value{
Data frame of treatment pairs with study information
}
\description{
Extracts all pairwise treatment comparisons from multi-arm studies
}
