% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{forest_plot_classic}
\alias{forest_plot_classic}
\title{Forest Plot - Fixed Version}
\usage{
forest_plot_classic(data, title = "Forest Plot", show_pooled = TRUE)
}
\arguments{
\item{data}{Data frame with effect_size, se, study columns}

\item{title}{Plot title}

\item{show_pooled}{Show pooled estimate}
}
\description{
Forest Plot - Fixed Version
}
