% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_ci}
\alias{format_ci}
\title{Format Confidence Intervals}
\usage{
format_ci(estimate, lower, upper, digits = 2)
}
\arguments{
\item{estimate}{Point estimate}

\item{lower}{Lower bound}

\item{upper}{Upper bound}

\item{digits}{Number of decimal places}
}
\value{
Formatted CI string
}
\description{
Format Confidence Intervals
}
