% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{funnel_plot_classic}
\alias{funnel_plot_classic}
\title{Funnel Plot - Fixed Version}
\usage{
funnel_plot_classic(data, title = "Funnel Plot", show_funnel = TRUE)
}
\arguments{
\item{data}{Data frame with effect_size and se columns}

\item{title}{Plot title}

\item{show_funnel}{Show funnel boundaries}
}
\description{
Funnel Plot - Fixed Version
}
