% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{funnel_plot_contour}
\alias{funnel_plot_contour}
\title{Contour Funnel Plot - Fixed Version}
\usage{
funnel_plot_contour(data, title = "Contour-Enhanced Funnel Plot")
}
\arguments{
\item{data}{Data frame with effect_size and se columns}

\item{title}{Plot title}
}
\description{
Contour Funnel Plot - Fixed Version
}
