% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novelty.R
\name{generate_advisor_recommendations}
\alias{generate_advisor_recommendations}
\title{Generate Advisor Recommendations}
\usage{
generate_advisor_recommendations(results, config)
}
\arguments{
\item{results}{Full CBAMM results object}

\item{config}{CBAMM configuration}
}
\value{
List of recommendations
}
\description{
Generate Advisor Recommendations
}
\keyword{internal}
