% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null Coalescing Operator.}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Second value}
}
\value{
First non-NULL value
}
\description{
Returns first non-NULL value
}
\keyword{internal}
