% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_meta_analysis.R
\name{is_network_connected}
\alias{is_network_connected}
\title{Check Network Connectivity}
\usage{
is_network_connected(treatment_pairs, treatments)
}
\arguments{
\item{treatment_pairs}{Data frame of treatment pairs}

\item{treatments}{Vector of all treatments}
}
\value{
Logical indicating if network is connected
}
\description{
Determines if all treatments are connected in the network
}
