% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_dashboard.R
\name{launch_cbamm_dashboard}
\alias{launch_cbamm_dashboard}
\title{Launch CBAMM Dashboard}
\usage{
launch_cbamm_dashboard()
}
\description{
Launches an interactive Shiny dashboard for meta-analysis
}
\examples{
\dontrun{
launch_cbamm_dashboard()
}
}
