% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unified_plotting.R
\name{launch_interactive_plots}
\alias{launch_interactive_plots}
\title{Interactive Plot Launcher}
\usage{
launch_interactive_plots(x)
}
\arguments{
\item{x}{CBAMM analysis object}
}
\value{
Interactive plot interface
}
\description{
Launches interactive plotting interface for CBAMM objects
}
