% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_meta_analysis.R
\name{ml_meta_analysis}
\alias{ml_meta_analysis}
\title{Machine Learning Meta-Analysis}
\usage{
ml_meta_analysis(data, method = "random_forest", ...)
}
\arguments{
\item{data}{Input data for analysis}

\item{method}{ML method to use (default: 'random_forest')}

\item{...}{Additional arguments}
}
\value{
ML meta-analysis results
}
\description{
Performs meta-analysis using machine learning techniques
}
