% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{p_curve_plot}
\alias{p_curve_plot}
\title{P-Curve Plot - Working Version}
\usage{
p_curve_plot(data, title = "P-Curve Analysis")
}
\arguments{
\item{data}{Data frame with effect_size and se columns}

\item{title}{Plot title}
}
\description{
P-Curve Plot - Working Version
}
