% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_meta_analysis.R
\name{plot.cbamm_cumulative}
\alias{plot.cbamm_cumulative}
\title{Plot Method for Cumulative Meta-Analysis}
\usage{
\method{plot}{cbamm_cumulative}(x, type = "estimate", ...)
}
\arguments{
\item{x}{cbamm_cumulative object}

\item{type}{Plot type: "estimate", "ci", "precision", "heterogeneity"}

\item{...}{Additional arguments}
}
\description{
Plot Method for Cumulative Meta-Analysis
}
