% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_meta_analysis.R
\name{plot.cbamm_network}
\alias{plot.cbamm_network}
\title{Plot Method for Network Meta-Analysis}
\usage{
\method{plot}{cbamm_network}(x, type = "forest", ...)
}
\arguments{
\item{x}{cbamm_network object}

\item{type}{Plot type: "network", "forest", "ranking"}

\item{...}{Additional arguments}
}
\description{
Plot Method for Network Meta-Analysis
}
