% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.cbamm_results}
\alias{plot.cbamm_results}
\title{Plot Method for CBAMM Results}
\usage{
\method{plot}{cbamm_results}(x, which = "all", ...)
}
\arguments{
\item{x}{CBAMM results object}

\item{which}{Which plot to show}

\item{...}{Additional arguments}
}
\value{
Plot object
}
\description{
Plot Method for CBAMM Results
}
