% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_meta_analysis.R
\name{plot_cumulative_ci}
\alias{plot_cumulative_ci}
\title{Plot Cumulative Confidence Intervals}
\usage{
plot_cumulative_ci(x, ...)
}
\arguments{
\item{x}{cbamm_cumulative object}

\item{...}{Additional arguments}
}
\description{
Plot Cumulative Confidence Intervals
}
