% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_meta_analysis.R
\name{plot_cumulative_estimates}
\alias{plot_cumulative_estimates}
\title{Plot Cumulative Estimates}
\usage{
plot_cumulative_estimates(x, ...)
}
\arguments{
\item{x}{cbamm_cumulative object}

\item{...}{Additional arguments}
}
\description{
Plot Cumulative Estimates
}
