% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_meta_analysis.R
\name{plot_cumulative_heterogeneity}
\alias{plot_cumulative_heterogeneity}
\title{Plot Cumulative Heterogeneity}
\usage{
plot_cumulative_heterogeneity(x, ...)
}
\arguments{
\item{x}{cbamm_cumulative object}

\item{...}{Additional arguments}
}
\description{
Plot Cumulative Heterogeneity
}
