% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_meta_analysis.R
\name{plot_cumulative_precision}
\alias{plot_cumulative_precision}
\title{Plot Cumulative Precision}
\usage{
plot_cumulative_precision(x, ...)
}
\arguments{
\item{x}{cbamm_cumulative object}

\item{...}{Additional arguments}
}
\description{
Plot Cumulative Precision
}
