% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_meta_analysis.R
\name{plot_forest_network}
\alias{plot_forest_network}
\title{Plot Forest Plot for Network Meta-Analysis}
\usage{
plot_forest_network(x, ...)
}
\arguments{
\item{x}{cbamm_network object}

\item{...}{Additional arguments}
}
\description{
Plot Forest Plot for Network Meta-Analysis
}
