% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_plotting.R
\name{plot_funnel_enhanced}
\alias{plot_funnel_enhanced}
\title{Enhanced Funnel Plot}
\usage{
plot_funnel_enhanced(
  ma_data,
  method = "standard",
  studies_to_highlight = NULL,
  color_by = NULL
)
}
\arguments{
\item{ma_data}{Meta-analysis data}

\item{method}{Funnel plot type: "standard", "contour", "trim_fill", "regression"}

\item{studies_to_highlight}{Vector of study IDs to highlight}

\item{color_by}{Variable to color points by}
}
\value{
ggplot object
}
\description{
Creates publication bias assessment plots with multiple enhancement options
}
