% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_plotting.R
\name{plot_inconsistency}
\alias{plot_inconsistency}
\title{Inconsistency Assessment Plot}
\usage{
plot_inconsistency(nma_results, method = "node_split")
}
\arguments{
\item{nma_results}{Network meta-analysis results with inconsistency assessment}

\item{method}{Inconsistency method: "node_split", "design_inconsistency", "loop"}
}
\value{
ggplot object
}
\description{
Visualizes network inconsistency through various methods
}
