% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_plotting.R
\name{plot_influence_analysis}
\alias{plot_influence_analysis}
\title{Influence Analysis Plot}
\usage{
plot_influence_analysis(ma_results, influence_measure = "cook_distance")
}
\arguments{
\item{ma_results}{Meta-analysis results}

\item{influence_measure}{Measure: "cook_distance", "leverage", "studentized"}
}
\value{
ggplot object
}
\description{
Visualizes influence of individual studies on pooled estimate
}
