% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_meta_analysis.R
\name{plot_network_diagram}
\alias{plot_network_diagram}
\title{Plot Network Diagram}
\usage{
plot_network_diagram(x, ...)
}
\arguments{
\item{x}{cbamm_network object}

\item{...}{Additional arguments}
}
\description{
Plot Network Diagram
}
