% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_plotting.R
\name{plot_outlier_detection}
\alias{plot_outlier_detection}
\title{Outlier Detection Plot}
\usage{
plot_outlier_detection(ma_data, method = "studentized")
}
\arguments{
\item{ma_data}{Meta-analysis data}

\item{method}{Outlier detection method}
}
\value{
ggplot object
}
\description{
Identifies and visualizes potential outliers in meta-analysis
}
