% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_plotting.R
\name{plot_publication_bias_suite}
\alias{plot_publication_bias_suite}
\title{Publication Bias Assessment Suite}
\usage{
plot_publication_bias_suite(
  ma_data,
  tests = c("funnel", "egger", "begg", "trim_fill")
)
}
\arguments{
\item{ma_data}{Meta-analysis data}

\item{tests}{Vector of tests to include}
}
\value{
List of ggplot objects
}
\description{
Comprehensive publication bias visualization
}
