% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_plotting.R
\name{plot_treatment_ranking}
\alias{plot_treatment_ranking}
\title{Treatment Ranking Plot}
\usage{
plot_treatment_ranking(nma_results, plot_type = "rankogram", treatments = NULL)
}
\arguments{
\item{nma_results}{Network meta-analysis results}

\item{plot_type}{Type: "rankogram", "cumulative", "sucra"}

\item{treatments}{Vector of treatments to include (default: all)}
}
\value{
ggplot object
}
\description{
Visualizes treatment rankings with uncertainty (rankograms or cumulative ranking plots)
}
