% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_meta_analysis.R
\name{plot_treatment_rankings}
\alias{plot_treatment_rankings}
\title{Plot Treatment Rankings}
\usage{
plot_treatment_rankings(x, ...)
}
\arguments{
\item{x}{cbamm_network object}

\item{...}{Additional arguments}
}
\description{
Plot Treatment Rankings
}
