% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_meta_analysis.R
\name{prepare_cumulative_data}
\alias{prepare_cumulative_data}
\title{Prepare Cumulative Data}
\usage{
prepare_cumulative_data(data, order_by, order_direction, custom_order)
}
\arguments{
\item{data}{Input data frame}

\item{order_by}{Ordering variable}

\item{order_direction}{Ordering direction}

\item{custom_order}{Custom ordering vector}
}
\value{
Ordered data frame
}
\description{
Orders data according to specified criteria for cumulative analysis
}
