% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_meta_analysis.R
\name{prepare_network_data}
\alias{prepare_network_data}
\title{Prepare Network Data Structure}
\usage{
prepare_network_data(data, studies, treatments, effect_size, std_error)
}
\arguments{
\item{data}{Input data frame}

\item{studies}{Study identifier column name}

\item{treatments}{Treatment identifier column name}

\item{effect_size}{Effect size column name}

\item{std_error}{Standard error column name}
}
\value{
Formatted network data structure
}
\description{
Converts input data to internal network format
}
