% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_suite.R
\name{print.cbamm_bayesian}
\alias{print.cbamm_bayesian}
\title{Print method for Bayesian meta-analysis}
\usage{
\method{print}{cbamm_bayesian}(x, ...)
}
\description{
Print method for Bayesian meta-analysis
}
