% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_meta_analysis.R
\name{print.cbamm_cumulative}
\alias{print.cbamm_cumulative}
\title{Print Method for Cumulative Meta-Analysis}
\usage{
\method{print}{cbamm_cumulative}(x, ...)
}
\arguments{
\item{x}{cbamm_cumulative object}

\item{...}{Additional arguments}
}
\description{
Print Method for Cumulative Meta-Analysis
}
