% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_methods.R, R/ipd_meta_analysis.R
\name{print.cbamm_ipd}
\alias{print.cbamm_ipd}
\title{Print method for IPD results}
\usage{
\method{print}{cbamm_ipd}(x, ...)

\method{print}{cbamm_ipd}(x, ...)
}
\description{
Print method for IPD results

Print method for IPD results
}
