% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_methods.R,
%   R/living_systematic_review.R
\name{print.cbamm_living}
\alias{print.cbamm_living}
\title{Print method for Living Review}
\usage{
\method{print}{cbamm_living}(x, ...)

\method{print}{cbamm_living}(x, ...)
}
\description{
Print method for Living Review

Print method for living review
}
