% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_enhanced.R
\name{print.cbamm_multivariate}
\alias{print.cbamm_multivariate}
\title{Print Multivariate Results}
\usage{
\method{print}{cbamm_multivariate}(x, digits = 4, ...)
}
\description{
Print Multivariate Results
}
