% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cbamm_results}
\alias{print.cbamm_results}
\title{S3 Methods for CBAMM Objects.}
\usage{
\method{print}{cbamm_results}(x, ...)
}
\arguments{
\item{x}{CBAMM results object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisible x
}
\description{
Print, summary, and plot methods for CBAMM results and configuration objects
Print Method for CBAMM Results
}
