% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_meta.R
\name{print.cbamm_timeseries}
\alias{print.cbamm_timeseries}
\title{Print method for time-series meta-analysis}
\usage{
\method{print}{cbamm_timeseries}(x, ...)
}
\description{
Print method for time-series meta-analysis
}
