% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_bias_toolkit.R
\name{publication_bias_suite}
\alias{publication_bias_suite}
\title{Publication Bias Suite}
\usage{
publication_bias_suite(ma_result)
}
\arguments{
\item{ma_result}{Meta-analysis result}
}
\description{
Publication Bias Suite
}
