% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_meta_analysis.R
\name{run_basic_random_effects}
\alias{run_basic_random_effects}
\title{Run Basic Random Effects Meta-Analysis}
\usage{
run_basic_random_effects(data)
}
\arguments{
\item{data}{Current data subset}
}
\value{
Meta-analysis results
}
\description{
Basic implementation of random effects meta-analysis
}
