% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{run_bayesian_analysis}
\alias{run_bayesian_analysis}
\title{Run Bayesian Meta-Analysis}
\usage{
run_bayesian_analysis(data, config)
}
\arguments{
\item{data}{Study data with weights}

\item{config}{CBAMM configuration}
}
\value{
List with Bayesian results
}
\description{
Run Bayesian Meta-Analysis
}
\keyword{internal}
