% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_meta_analysis.R
\name{run_bayesian_ma}
\alias{run_bayesian_ma}
\title{Run Bayesian Meta-Analysis}
\usage{
run_bayesian_ma(data)
}
\arguments{
\item{data}{Current data subset}
}
\value{
Meta-analysis results
}
\description{
Simple Bayesian implementation for cumulative analysis
}
