% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_meta_analysis.R
\name{run_frequentist_network_ma}
\alias{run_frequentist_network_ma}
\title{Run Frequentist Network Meta-Analysis}
\usage{
run_frequentist_network_ma(network_data, reference, model, method)
}
\arguments{
\item{network_data}{Prepared network data}

\item{reference}{Reference treatment}

\item{model}{Model type ("random" or "fixed")}

\item{method}{Estimation method ("REML" or "ML")}
}
\value{
Network meta-analysis results
}
\description{
Performs frequentist network meta-analysis using multivariate meta-analysis
}
