% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{run_publication_bias_sensitivity}
\alias{run_publication_bias_sensitivity}
\title{Run Publication Bias Sensitivity Analysis}
\usage{
run_publication_bias_sensitivity(data)
}
\arguments{
\item{data}{Study data}
}
\value{
List of publication bias test results
}
\description{
Run Publication Bias Sensitivity Analysis
}
\keyword{internal}
