% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_meta_analysis.R
\name{run_random_effects_ma}
\alias{run_random_effects_ma}
\title{Run Random Effects Meta-Analysis}
\usage{
run_random_effects_ma(data)
}
\arguments{
\item{data}{Current data subset}
}
\value{
Meta-analysis results
}
\description{
Performs random effects meta-analysis for current data subset
}
