% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{run_stratified_analysis}
\alias{run_stratified_analysis}
\title{Run Stratified Meta-Analysis}
\usage{
run_stratified_analysis(data, config)
}
\arguments{
\item{data}{Study data}

\item{config}{CBAMM configuration}
}
\value{
List of stratified results
}
\description{
Run Stratified Meta-Analysis
}
\keyword{internal}
