% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{rve_print}
\alias{rve_print}
\title{Run RVE Analysis and Print Results}
\usage{
rve_print(res, cluster_vec, label = "[RVE-CR2]")
}
\arguments{
\item{res}{Meta-analysis result}

\item{cluster_vec}{Cluster vector for RVE}

\item{label}{Label for output}
}
\value{
RVE results (invisibly)
}
\description{
Run RVE Analysis and Print Results
}
\keyword{internal}
