% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{safe_run_core_meta_analysis}
\alias{safe_run_core_meta_analysis}
\title{Safe Meta-Analysis Runner}
\usage{
safe_run_core_meta_analysis(data, config)
}
\description{
Safe Meta-Analysis Runner
}
\keyword{internal}
