% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_methods.R, R/diagnostic_accuracy.R
\name{simulate_dta_data}
\alias{simulate_dta_data}
\title{Simulate Diagnostic Test Data}
\usage{
simulate_dta_data(
  n_studies = 10,
  sens_mean = 0.8,
  spec_mean = 0.9,
  n_diseased = 100,
  n_healthy = 100
)

simulate_dta_data(
  n_studies = 10,
  sens_mean = 0.8,
  spec_mean = 0.9,
  n_diseased = 100,
  n_healthy = 100
)
}
\arguments{
\item{n_studies}{Number of studies to simulate}

\item{sens_mean}{Mean sensitivity}

\item{spec_mean}{Mean specificity}

\item{n_diseased}{Number of diseased patients per study}

\item{n_healthy}{Number of healthy patients per study}
}
\value{
Data frame with diagnostic test data
}
\description{
Simulate Diagnostic Test Data

Simulate Diagnostic Test Accuracy Data
}
