% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_methods.R, R/ipd_meta_analysis.R
\name{simulate_ipd_data}
\alias{simulate_ipd_data}
\title{Simulate IPD Data}
\usage{
simulate_ipd_data(
  n_studies = 5,
  n_patients = 100,
  treatment_effect = 0.5,
  between_study_sd = 0.2
)

simulate_ipd_data(
  n_studies = 5,
  n_patients = 100,
  treatment_effect = 0.5,
  between_study_sd = 0.2
)
}
\arguments{
\item{n_studies}{Number of studies}

\item{n_patients}{Number of patients per study}

\item{treatment_effect}{True treatment effect}

\item{between_study_sd}{Between-study standard deviation}
}
\value{
List of data frames
}
\description{
Simulate IPD Data

Simulate IPD Data
}
