% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_methods.R,
%   R/living_systematic_review.R
\name{simulate_sequential_data}
\alias{simulate_sequential_data}
\title{Simulate Sequential Data}
\usage{
simulate_sequential_data(n_studies = 10, true_effect = 0.3, start_id = 1)

simulate_sequential_data(n_studies = 10, true_effect = 0.3, start_id = 1)
}
\arguments{
\item{n_studies}{Number of studies}

\item{true_effect}{True effect size}

\item{start_id}{Starting study ID}
}
\value{
Data frame with sequential data
}
\description{
Simulate Sequential Data

Simulate Sequential Data
}
