% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_meta_analysis.R
\name{summary.cbamm_cumulative}
\alias{summary.cbamm_cumulative}
\title{Summary Method for Cumulative Meta-Analysis}
\usage{
\method{summary}{cbamm_cumulative}(object, ...)
}
\arguments{
\item{object}{cbamm_cumulative object}

\item{...}{Additional arguments}
}
\description{
Summary Method for Cumulative Meta-Analysis
}
