% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/living_systematic_review.R
\name{summary.cbamm_living}
\alias{summary.cbamm_living}
\title{Summary method for living review}
\usage{
\method{summary}{cbamm_living}(object, ...)
}
\description{
Summary method for living review
}
