% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_meta_analysis.R
\name{summary.cbamm_network}
\alias{summary.cbamm_network}
\title{Summary Method for Network Meta-Analysis}
\usage{
\method{summary}{cbamm_network}(object, ...)
}
\arguments{
\item{object}{cbamm_network object}

\item{...}{Additional arguments}
}
\description{
Summary Method for Network Meta-Analysis
}
