% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.cbamm_results}
\alias{summary.cbamm_results}
\title{Summary Method for CBAMM Results}
\usage{
\method{summary}{cbamm_results}(object, ...)
}
\arguments{
\item{object}{CBAMM results object}

\item{...}{Additional arguments (unused)}
}
\value{
Summary data frame
}
\description{
Summary Method for CBAMM Results
}
