% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_meta.R
\name{time_series_meta}
\alias{time_series_meta}
\title{Time-Series Meta-Analysis}
\usage{
time_series_meta(
  data,
  time_var = "time",
  effect_var = "effect",
  study_var = "study",
  method = c("multilevel", "growth_curve")
)
}
\arguments{
\item{data}{Data with multiple time points per study}

\item{time_var}{Name of time variable}

\item{effect_var}{Name of effect variable}

\item{study_var}{Name of study variable}

\item{method}{Analysis method}
}
\description{
Time-Series Meta-Analysis
}
