% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/living_systematic_review.R
\name{update_living_review}
\alias{update_living_review}
\title{Update Living Review}
\usage{
update_living_review(living_review, new_data, ...)
}
\arguments{
\item{living_review}{Existing living review object}

\item{new_data}{New data to add}

\item{...}{Additional arguments}
}
\value{
Updated living review object
}
\description{
Update Living Review
}
