% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_config}
\alias{validate_config}
\title{Validate Package configuration settings..}
\usage{
validate_config(
  methods,
  estimators,
  output,
  transportability,
  bayesian,
  sensitivity
)
}
\description{
Internal function to validate configuration parameters
}
\keyword{internal}
