% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_meta_analysis.R
\name{validate_cumulative_inputs}
\alias{validate_cumulative_inputs}
\title{Validate Cumulative Meta-Analysis Inputs}
\usage{
validate_cumulative_inputs(data, order_by, order_direction, method)
}
\arguments{
\item{data}{Input data frame}

\item{order_by}{Ordering variable}

\item{order_direction}{Ordering direction}

\item{method}{Analysis method}
}
\description{
Validates input parameters for cumulative meta-analysis
}
