% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_meta_analysis.R
\name{validate_network_inputs}
\alias{validate_network_inputs}
\title{Validate Network Meta-Analysis Inputs}
\usage{
validate_network_inputs(data, studies, treatments, effect_size, std_error)
}
\arguments{
\item{data}{Input data frame}

\item{studies}{Study identifier column name}

\item{treatments}{Treatment identifier column name}

\item{effect_size}{Effect size column name}

\item{std_error}{Standard error column name}
}
\description{
Validates input parameters for network meta-analysis
}
