% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_meta_analysis.R
\name{validate_network_structure}
\alias{validate_network_structure}
\title{Validate Network Structure}
\usage{
validate_network_structure(data, studies, treatments)
}
\arguments{
\item{data}{Input data frame}

\item{studies}{Study identifier column name}

\item{treatments}{Treatment identifier column name}
}
\description{
Ensures the data forms a connected network
}
