% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_fallback}
\alias{with_fallback}
\title{Safe Package Loading with Fallback.}
\usage{
with_fallback(pkg, fun_call, fallback_fun = NULL, fallback_msg = NULL)
}
\arguments{
\item{pkg}{Package name}

\item{fun_call}{Function call to attempt}

\item{fallback_fun}{Fallback function if package unavailable}

\item{fallback_msg}{Message to show when using fallback}
}
\description{
Attempt to use a package function with fallback behavior
}
\keyword{internal}
